+++ Kernel/Arch/riscv64/Timer.cpp
@@ -9,6 +9,7 @@
 #include <Kernel/Arch/riscv64/CPU.h>
 #include <Kernel/Arch/riscv64/SBI.h>
 #include <Kernel/Arch/riscv64/Timer.h>
+#include <Kernel/Firmware/Devicetree/Devicetree.h>
 
 namespace Kernel::RISCV64 {
